# Conflicts:
#	browserassets/tgui/tgui.bundle.js
#	code/WorkInProgress/artemis/GeneratePlanets.dm
#	code/WorkInProgress/artemis/planetary_teleporter.dm
#	code/modules/worldgen/prefab/planet.dm
#	goonstation.dme

+++ code/procs/jobprocs.dm
@@ -710,12 +710,13 @@ var/global/totally_random_jobs = FALSE
 		if (src.traitHolder && src.traitHolder.hasTrait("onearmed"))
 			if (src.limbs)
 				SPAWN(6 SECONDS)
-					if (prob(50))
-						if (src.limbs.l_arm)
-							qdel(src.limbs.l_arm.remove(0))
-					else
-						if (src.limbs.r_arm)
-							qdel(src.limbs.r_arm.remove(0))
+					if (ishuman(src))
+						if (prob(50))
+							if (src.limbs.l_arm)
+								qdel(src.limbs.l_arm.remove(0))
+						else
+							if (src.limbs.r_arm)
+								qdel(src.limbs.r_arm.remove(0))
 					boutput(src, "<b>Your singular arm makes you feel responsible for crimes you couldn't possibly have committed.</b>" )
 
 		if (src.traitHolder && src.traitHolder.hasTrait("nolegs"))
